Full factorial changes
the full factorial now generates a table with the headers as factors and the data the high and low levels in real life, not just -1 and +1. The input is now a dictionary of the factors and their low and high levels, e.g. factors = {'Temp':[30,37],'Enzyme':['A,'B']} This can currently only deal with two level designs.